+++ crates/miden-objects/src/account/mod.rs
@@ -17,6 +17,7 @@ pub use account_id::{
     AccountIdVersion,
     AccountStorageMode,
     AccountType,
+    CustomNetworkId,
     NetworkId,
 };